homework_project/exadel_project/requirements.txt
amqp==5.1.1
asgiref==3.5.1
attrs==21.4.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
drf-yasg==1.20.0
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2

